ground unit grave yard
I think I added all ground units to the graveyard, they are invisible and invincible, and unmovable. I hope it helps, if it does, I can add the airplanes to it as well, don't know how to add the FC3 though, but 45 minutes into the mission and so far so good.